Squashing warnings.

+++ src/libambulant/common/region.cpp
@@ -630,7 +630,7 @@ surface_impl::get_fit_rect(const lib::size& src_size, lib::rect* out_src_rect, c
 		x_region_for_image_left = (int)((-x_image_for_region_left * scale_horizontal) + 0.5);
 		x_image_for_region_left = 0;
 	}
-	if (x_image_for_region_right > src_real_size.w) {
+	if (x_image_for_region_right > (int)src_real_size.w) {
 		int overshoot = x_image_for_region_right - src_real_size.w;
 		x_image_for_region_right = src_real_size.w;
 		x_region_for_image_right = x_region_for_image_right - (int)((overshoot * scale_horizontal) + 0.5);
@@ -639,7 +639,7 @@ surface_impl::get_fit_rect(const lib::size& src_size, lib::rect* out_src_rect, c
 		y_region_for_image_top = (int)((-y_image_for_region_top * scale_horizontal) + 0.5);
 		y_image_for_region_top = 0;
 	}
-	if (y_image_for_region_bottom > src_real_size.h) {
+	if (y_image_for_region_bottom > (int)src_real_size.h) {
 		int overshoot = y_image_for_region_bottom - src_real_size.h;
 		y_image_for_region_bottom = src_real_size.h;
 		y_region_for_image_bottom = y_region_for_image_bottom - (int)((overshoot * scale_vertical) + 0.5);
